+++ multibody/inverse_kinematics/BUILD.bazel
@@ -19,12 +19,12 @@ drake_cc_package_library(
     deps = [
         ":global_inverse_kinematics",
         ":inverse_kinematics_core",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
     ],
 )
 
 drake_cc_library(
-    name = "kinematic_constraint",
+    name = "kinematic_evaluators",
     srcs = [
         "angle_between_vectors_constraint.cc",
         "com_in_polyhedron_constraint.cc",
@@ -32,7 +32,7 @@ drake_cc_library(
         "distance_constraint.cc",
         "distance_constraint_utilities.cc",
         "gaze_target_constraint.cc",
-        "kinematic_constraint_utilities.cc",
+        "kinematic_evaluator_utilities.cc",
         "minimum_distance_constraint.cc",
         "orientation_constraint.cc",
         "point_to_point_distance_constraint.cc",
@@ -46,7 +46,7 @@ drake_cc_library(
         "distance_constraint.h",
         "distance_constraint_utilities.h",
         "gaze_target_constraint.h",
-        "kinematic_constraint_utilities.h",
+        "kinematic_evaluator_utilities.h",
         "minimum_distance_constraint.h",
         "orientation_constraint.h",
         "point_to_point_distance_constraint.h",
@@ -74,7 +74,7 @@ drake_cc_library(
     ],
     visibility = ["//visibility:private"],
     deps = [
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//multibody/plant",
         "//solvers:mathematical_program",
     ],
@@ -125,7 +125,7 @@ drake_cc_googletest(
     name = "position_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
     ],
 )
 
@@ -133,7 +133,7 @@ drake_cc_googletest(
     name = "com_position_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
     ],
 )
 
@@ -141,7 +141,7 @@ drake_cc_googletest(
     name = "com_in_polyhedron_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//common/test_utilities:expect_throws_message",
     ],
 )
@@ -150,7 +150,7 @@ drake_cc_googletest(
     name = "orientation_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//math:geometric_transform",
     ],
 )
@@ -159,7 +159,7 @@ drake_cc_googletest(
     name = "gaze_target_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
     ],
 )
 
@@ -167,7 +167,7 @@ drake_cc_googletest(
     name = "angle_between_vectors_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
     ],
 )
 
@@ -175,7 +175,7 @@ drake_cc_googletest(
     name = "minimum_distance_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//common/test_utilities:expect_throws_message",
         "//solvers/test_utilities",
     ],
@@ -185,7 +185,7 @@ drake_cc_googletest(
     name = "distance_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//solvers/test_utilities",
     ],
 )
@@ -194,7 +194,7 @@ drake_cc_googletest(
     name = "point_to_point_distance_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//solvers/test_utilities",
     ],
 )
@@ -203,7 +203,7 @@ drake_cc_googletest(
     name = "unit_quaternion_constraint_test",
     deps = [
         ":inverse_kinematics_test_utilities",
-        ":kinematic_constraint",
+        ":kinematic_evaluators",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:symbolic_test_util",
         "//solvers:solve",
