+++ multibody/inverse_kinematics/distance_constraint.cc
@@ -5,7 +5,7 @@
 #include <vector>
 
 #include "drake/multibody/inverse_kinematics/distance_constraint_utilities.h"
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
