+++ multibody/inverse_kinematics/kinematic_evaluator_utilities.cc
@@ -1,4 +1,4 @@
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
