+++ multibody/inverse_kinematics/minimum_distance_constraint.cc
@@ -6,7 +6,7 @@
 #include <Eigen/Dense>
 
 #include "drake/multibody/inverse_kinematics/distance_constraint_utilities.h"
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
