+++ multibody/inverse_kinematics/orientation_constraint.cc
@@ -1,7 +1,7 @@
 #include "drake/multibody/inverse_kinematics/orientation_constraint.h"
 
 #include "drake/math/autodiff_gradient.h"
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 using drake::multibody::internal::RefFromPtrOrThrow;
 using drake::multibody::internal::UpdateContextConfiguration;
