+++ multibody/optimization/manipulator_equation_constraint.cc
@@ -2,7 +2,7 @@
 
 #include <unordered_map>
 
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
