+++ multibody/optimization/static_friction_cone_complementarity_constraint.cc
@@ -2,7 +2,7 @@
 
 #include <limits>
 
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
