+++ multibody/optimization/static_friction_cone_constraint.cc
@@ -3,7 +3,7 @@
 #include <limits>
 #include <vector>
 
-#include "drake/multibody/inverse_kinematics/kinematic_constraint_utilities.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
